# api/openapi.yaml
openapi: 3.1.0
info:
  title: HEOS control service
  version: 0.9.3
  license:
    name: MIT
    identifier: MIT
  description: |
    Observe configured HEOS players, browse allowed local media, and submit
    durable playback operations with optional volume ramps and automatic stop.

    ## Authentication
    Enter your machine token under **Authentication → Bearer Token**. Requests use
    `Authorization: Bearer <token>`. Scopes are additive: `read` allows
    observation and preflight, `control` allows direct volume/mute/transport,
    and `operator` allows playback and operator stop. Every request also needs
    a grant for its player; `operator` does not imply `read` or `control`.
    Health probes are public. Device writes require configured per-player opt-in,
    verified identity and volume policy, and fresh state.

    ## Playback workflow
    1. List players and sources, then browse items to obtain a current `item_ref`.
       References are opaque; do not construct them from names or device IDs.
    2. Send the concrete playback body to **Preflight**. This read-only check
       changes neither playback nor the journal and does not reserve a player.
    3. Submit the same body to **Start playback** with a stable `Idempotency-Key`
       and the preflight player's revision in a quoted `If-Match` header.
    4. Follow `Location` or read `/v1/operations/{operation}` to observe execution.
...

    ## Retries and state
...
servers:
- url: /
...
paths:
...
  /v1/players:
    get:
      operationId: getPlayers
      tags: [Players]
      summary: List authorized players
      description: Requires read scope. Returns configured players permitted by the credential, including freshness and availability. Unknown or stale state is not a write grant.
      parameters: []
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/players/{player}:
    get:
      operationId: getPlayer
      tags: [Players]
      summary: Read player state
      description: Requires read scope and a grant for this player. Returns an observed snapshot and ETag. Quote the revision in If-Match for a new mutation; admission independently revalidates it.
      parameters:
      - &id001
        name: player
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,63}$
          maxLength: 64
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
            ETag:
              description: Quoted state revision including process epoch.
              schema:
                type: string
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/groups:
    get:
      operationId: getGroups
      tags: [Players]
      summary: Read group membership
      description: Requires read scope. Projects only authorized configured players; this is not a complete inventory of every room. Grouped targets are not writable.
      parameters: []
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/sources:
    get:
      operationId: getSources
      tags: [Sources]
      summary: List local media sources
      description: Requires read scope. Lists configured sources for authorized players; availability depends on current HEOS discovery.
      parameters: []
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Source'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/sources/{source}/items:
    get:
      operationId: getItems
      tags: [Sources]
      summary: Browse source items
      description: Requires read scope and access to the source player. Pass a browsable item_ref as parent_ref to enter a container, and next_cursor as cursor to continue a listing, with bounded traversal. Use an available playable item_ref for preflight and playback.
      parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,63}$
          maxLength: 64
      - name: parent_ref
        in: query
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 256
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 256
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Items'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/players/{player}/queue:
    get:
      operationId: getQueue
      tags: [Players]
      summary: Read a queue page
      description: Requires read scope and a grant for this player. Follow next_offset while it is non-null. A stale or partial queue is not evidence of playback ownership.
      parameters:
      - name: player
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,63}$
          maxLength: 64
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          maximum: 10000
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
      - name: revision
        in: query
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 160
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/players/{player}/preflight:
    post:
      operationId: preflight
      tags: [Playback]
      summary: Check playback readiness
      description: Read-only check of a concrete PlaybackInput, including bounded media resolution. Returns readiness warnings and an observed player revision; does not reserve the player, persist a run, select media or change playback. No Idempotency-Key or If-Match is required. Read scope and player ACL apply; takeover requires operator scope. Playback revalidates independently.
      parameters:
      - name: player
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,63}$
          maxLength: 64
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preflight'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackInput'
...
  /v1/operations/{operation}:
    get:
      operationId: getOperation
      tags: [Operations]
      summary: Read operation progress and result
      description: Requires read scope and a player grant. The creator may read its operation; operator scope additionally permits other creators on the same authorized player. Acceptance is not success; inspect state, phase, error_code and outcome.
      parameters:
      - &id002
        name: operation
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
      responses:
        '200':
          description: Current read-only result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/events:
    get:
      operationId: getEvents
      tags: [Events]
      summary: Subscribe to change notifications
      parameters:
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 160
      responses:
        '200':
          description: Current read-only result
          content:
            text/event-stream:
              schema:
                type: string
              example: 'id: epoch:1

                event: snapshot_required

                data: {}


                '
          headers:
            X-Request-ID:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          description: Rejected request; no playback command was sent.
          headers:
            X-Request-ID:
              schema:
                type: string
            WWW-Authenticate:
              schema:
                type: string
            Retry-After:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description: 'Bounded replay. Event kinds: player_changed, operation_changed,
        snapshot_required. Data contains authorized player and optional operation
        identifiers only; fetch snapshots after a gap/restart. Slow clients disconnect.
        Heartbeat every 15 seconds.'
  /v1/players/{player}/playback:
    post:
      operationId: playback
      tags: [Playback]
      summary: Start playback
      description: Requires operator scope.
      parameters:
      - *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 256
          pattern: ^[A-Za-z0-9._:-]+$
      - name: If-Match
        in: header
        required: false
        description: Required for a new request (428 if absent); replay precedes freshness
          checks.
        schema:
          type: string
          minLength: 1
          maxLength: 256
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackInput'
      responses:
        '202':
          description: Durably admitted operation or current idempotent replay; acceptance
            is not device confirmation.
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        default:
          description: Request error. Inspect the same idempotency key after an uncertain
            commit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/players/{player}/stop:
    post:
      operationId: stopPlayer
      tags: [Playback]
      summary: Stop playback as operator
      description: Requires operator scope.
      parameters:
      - *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 256
          pattern: ^[A-Za-z0-9._:-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopInput'
      responses:
        '202':
          description: Durably admitted operation or current idempotent replay; acceptance
            is not device confirmation.
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        default:
          description: Request error. Inspect the same idempotency key after an uncertain
            commit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/operations/{operation}/cancel:
    post:
      operationId: cancelOperation
      tags: [Operations]
      summary: Cancel an operation
      description: Requires control scope. Takeover additionally requires operator
        scope.
      parameters:
      - *id002
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 256
          pattern: ^[A-Za-z0-9._:-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelInput'
      responses:
        '202':
          description: Durably admitted operation or current idempotent replay; acceptance
            is not device confirmation.
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        default:
          description: Request error. Inspect the same idempotency key after an uncertain
            commit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    machineToken:
      type: http
      scheme: bearer
      description: Machine token over verified HTTPS, explicit route scope and player
        allowlist. Read does not imply control; operator does not imply read.
  schemas:
    Player:
      type: object
      additionalProperties: false
      required:
      - key
      - availability
      - observed_at
      - stale
      - revision
      - playback_state
      - volume
      - muted
      - grouped
      - capabilities
      - active_operation
      properties:
        key:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,63}$
          maxLength: 64
        availability:
          type: string
          enum:
          - online
          - offline
          - unknown
        observed_at:
          type:
          - string
          - 'null'
          format: date-time
        stale:
          type: boolean
        revision:
          type: string
          maxLength: 160
        playback_state:
          type: string
          enum:
          - unknown
          - play
          - pause
          - stop
        volume:
          type: object
          additionalProperties: false
          required:
          - unit
          - level
          properties:
            unit:
              type: string
              const: heos
            level:
              type:
              - integer
              - 'null'
              minimum: 0
              maximum: 100
        muted:
          type:
          - boolean
          - 'null'
        grouped:
          type:
          - boolean
          - 'null'
        capabilities:
          type: object
          additionalProperties: false
          required:
          - read
          - playback
          - volume
          properties:
            read:
              type: string
              enum:
              - supported
              - unsupported
              - unverified
            playback:
              type: string
              enum:
              - supported
              - unsupported
              - unverified
            volume:
              type: string
              enum:
              - supported
              - unsupported
              - unverified
        active_operation:
          type:
          - string
          - 'null'
      x-go-type: control.Player
      x-go-type-import:
        path: github.com/dreylark/heos-control/internal/control
        name: control
...
security:
- machineToken: []